\name{surrosurvROC-package}
\alias{surrosurvROC-package}
\docType{package}
\title{Surrogate Survival ROC}
\description{Nonparametric and semiparametric estimations of the time-dependent ROC curve for an incomplete failure time data with surrogate failure time endpoints.}
\details{
\tabular{ll}{
Package: \tab isoph\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2018-08-17\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [cre]

Maintainer: Yunro Chung <yunro.chung@asu.edu>
}
\references{Yunro Chung and Yingye Zheng, Improving efficiency of evaluating prognostic accuracy of biomarkers for incomplete failure-time data with surrogate outcome (in progress)}
\keyword{biomarkers, inverse probability weighing, bootstrap}
